pyproject.toml
[project]
name = "mail_juicer"
version = "0.1.0"
description = "Manage mail from your server."
authors = [
    {name = "Mathias Millet", email="mathias@mmill.eu"},
]
license = {file = "LICENSE.txt"}
readme = "README.md"
requires-python = ">=3.6"

dependencies = [
  "typer>=0.7, <0.8",
  "pydantic>=1.10,<2",
  "pyyaml",
  "imap-tools"
]


[build-system]
requires = [
    "setuptools >= 60.9.3",
    "setuptools_scm[toml] >= 6.2"
]
build-backend = "setuptools.build_meta"


[project.optional-dependencies]
dev = [
    "pytest >= 7.0",
    "pre-commit >= 2.0",
]

[tool.pytest.ini_options]
background-cmd="python tests/imap_server.py"


[tool.pycln]
all = true

[tool.black]
line-length = 99

[tool.isort]
line_length = 99
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true